+++ src/app/models/test-result-status.ts
@@ -1,0 +1,5 @@
+export enum TestResultStatus {
+  Failed = 'Failed',
+  Passed = 'Passed',
+  Skipped = 'Skipped',
+}